+++ components/singleTarotCardWithFlip/SingleTarotCardWithFlip.tsx
@@ -1,5 +1,4 @@
 import { Box } from '@mui/material';
-import Image from 'next/image';
 
 interface SingleTarotCardWithFlipAction {
   onClick?: () => void;
@@ -61,7 +60,6 @@ const SingleTarotCard = ({
           borderRadius: 3,
           boxShadow: { boxShadow },
           backfaceVisibility: 'hidden',
-          overflow: 'hidden',
           "&[data-size='medium']": {
             width: '200px'
           },
@@ -69,26 +67,17 @@ const SingleTarotCard = ({
             width: '75px'
           }
         }}
-      >
-        <Image
-          src='/image/welcome-cards/tarot-card-back.webp'
-          alt='The tarot card back'
-          width={size === 'small' ? 75 : 200}
-          height={0}
-          style={{
-            width: '100%',
-            height: 'auto',
-            borderRadius: '12px'
-          }}
-        />
-      </Box>
+        component='img'
+        src='/image/welcome-cards/tarot-card-back.webp'
+        alt='The tarot card back'
+      />
       <Box
         data-size={size}
         sx={{
-          position: 'absolute',
           borderRadius: 3,
           boxShadow: { boxShadow },
           transformOrigin: 'center center',
+          position: 'absolute',
           backfaceVisibility: 'hidden',
           transform: 'rotateY(180deg)',
           "&[data-size='medium']": {
@@ -98,19 +87,10 @@ const SingleTarotCard = ({
             width: '100px'
           }
         }}
-      >
-        <Image
-          src={image}
-          alt={altText || ''}
-          width={size === 'small' ? 100 : 200}
-          height={0}
-          style={{
-            width: '100%',
-            height: 'auto',
-            borderRadius: '12px'
-          }}
-        />
-      </Box>
+        component='img'
+        src={image}
+        alt={altText}
+      />
     </Box>
   );
 };